Fix viewport for retina displays
On retina displays, framebuffer extent is twice the screen extent. To
account for this, the glViewport calls and post-processor framebuffers
must use the framebuffer extent instead of the GLFW window extent.

See https://www.glfw.org/faq.html#43---why-is-my-output-in-the-lower-left-corner-of-the-window
for details.

This is now fixed. We add the Extent struct and implement
Window::getFramebufferSize as utility.