Fix/macos (#5)
* Use forward-compatible core profile on macOS

Fixes "NSGL: The targeted version of macOS only supports
forward-compatible core profile contexts for OpenGL 3.2 and above"
error. See
https://www.glfw.org/faq.html#41---how-do-i-create-an-opengl-30-context
for details.

* Fix viewport for retina displays

On retina displays, framebuffer extent is twice the screen extent. To
account for this, the glViewport calls and post-processor framebuffers
must use the framebuffer extent instead of the GLFW window extent.

See https://www.glfw.org/faq.html#43---why-is-my-output-in-the-lower-left-corner-of-the-window
for details.

This is now fixed. We add the Extent struct and implement
Window::getFramebufferSize as utility.

* Fix moving game window between UHD displays

Moving the window between UHD displays with varying scaling will result
in wrong rendering. This is fixed now:

* PostProcessor now has a resize() method for updating the FBOs/textures
  according to the size of the default FBO.
* The Window class now has a resize callback which triggers the
  PostProcessor's resize method
* Rendering the main scene now happens in _all_ game states, as resizing
  during GAME_WIN, GAME_LOSE will otherwise result in losing the main
  scene during PostProcessor resizing.
* The user may still not resize the window, as collision detection
  doesn't work during resizing. In fact, glfwPollEvents blocks the game
  loop during user resizing, but delta time measurements remain the
  same. This result in the ball unexpectedly moving far out of frame.

* Add version of windows binary to README

* Fix erroneous resizing under Windows

Under windows, using two displays with different content scaling results
in incorrect resizing when moving the window between displays. This is
solved by updating the external dependency GLFW to 3.3.2 (most likely
the problem is caused by a bug in GLFW that is fixed somewhere between
3.0.0 and 3.3.2).